refactor(remix_api): split tests into seperate files
